dius: partial support for SyncFile on Iros
